Modularizes analysis workflow and enhances SAST scanning
Extracts the analysis workflow and related helpers into a dedicated module to improve code organization and readability. Refactors SAST analyzer to enable hybrid scanning with regex and partial scans for large or minified files, introduces new Semgrep rules for Chrome extension risks, and improves file filtering logic. Updates hashing utility for better validation and chunk handling. Adds Jinja2 to dependencies for future templating needs. Comments out legacy cache checks to streamline scan triggering.

Improves maintainability, extensibility, and scanning reliability.
